# src/main/resources/static/openapi.yaml
openapi: 3.0.1
info:
  title: Social-Media-Api
  version: v1
  description: "Спецификация для проекта Social-Media-Api."
servers:
  - url: http://localhost:8080
    description: local server
paths:
  /api/v1/post/get:
    get:
      security:
        - bearerAuth: [ ]
      tags:
        - Post endpoints
      summary: "Получение поста по id"
      operationId: getPost
      parameters:
        - name: postId
          description: "ID поста"
          in: query
          required: true
          allowEmptyValue: false
          example: 1
          schema:
            type: integer
      responses:
        '200':
          description: Post successful get
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostDto'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/post/getFeedUser:
    get:
      security:
        - bearerAuth: [ ]
      tags:
        - Post endpoints
      summary: "Получение постов (ленты активности) от пользователей на которых подписан пользователь"
      operationId: getFeedUser
      parameters:
        - name: page
          description: "Номер отображаемой страницы ленты активности (начинается с 0)"
          in: query
          required: true
          allowEmptyValue: false
          schema:
            type: integer
            default: 0
        - name: size
          description: "Размер отображаемой страницы ленты активности (начинается с 0)"
          in: query
          required: true
          allowEmptyValue: false
          schema:
            type: integer
            default: 2
        - name: isSortAsDesk
          description: "Сортировка по времени публикации постов (true - по убыванию, false - по возрастанию)"
          in: query
          required: true
          allowEmptyValue: false
          schema:
            type: boolean
            default: true
      responses:
        '200':
          description: Successful get last posts followers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PostDto'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Not forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/post/create:
    post:
      security:
        - bearerAuth: [ ]
      tags:
        - Post endpoints
      summary: "Создание поста"
      operationId: cretePost
      requestBody:
        description: "Ввод описания, сообщения и загрузка картинки для создания поста"
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                postDto:
                  $ref: '#/components/schemas/PostDto'
                file:
                  type: string
                  format: binary
      responses:
        '201':
          description: Post successful created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostDto'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/post/update:
    put:
      security:
        - bearerAuth: [ ]
      tags:
        - Post endpoints
      summary: "Изменение данных в посте"
      operationId: updatePost
      parameters:
        - name: postId
          description: "ID поста"
          in: query
          required: true
          allowEmptyValue: false
          example: 1
          schema:
            type: integer
      requestBody:
        description: "Ввод описания, сообщения и загрузка картинки для изменения существующего поста"
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                postDto:
                  $ref: '#/components/schemas/PostDto'
                file:
                  type: string
                  format: binary
      responses:
        '201':
          description: Post successful changed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostDto'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/post/getAllPosts:
    get:
      security:
        - bearerAuth: [ ]
      tags:
        - Post endpoints
      summary: "Получение всех постов из базы данных (Доступен только авторизованным пользователям с ролью ADMIN)"
      operationId: getAllPosts
      responses:
        '200':
          description: Successful get all posts from BD
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PostDto'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Not forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/post/delete:
    delete:
      security:
        - bearerAuth: [ ]
      tags:
        - Post endpoints
      summary: "Удаление поста по id"
      operationId: deletePost
      parameters:
        - name: postId
          description: "ID поста"
          in: query
          required: true
          allowEmptyValue: false
          example: 1
          schema:
            type: integer
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: string
                example: "Post with id = 1 deleted!"
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    PostDto:
      required:
        - description
        - message
      type: object
      properties:
        id:
          type: integer
          readOnly: true
          description: "Id поста"
          example: "1"
        description:
          type: string
          description: "Описание поста"
          example: "Description"
        message:
          type: string
          description: "Сообщение поста"
          example: "Message"
        imageLink:
          type: string
          readOnly: true
          description: "Ссылка на загруженную картинку"
          example: "http://localhost:9000/images/325c0226-8acb-41ad-a214-40ff773b35bc%3A%20%3C2024-01-11T16%3A35%3A05.852513900%3E%20image_name.jpeg?X-Amz-Algorithm=AWS4-HMAC-SHA256&X-Amz-Credential=minioadmin%2F20240111%2Fus-east-1%2Fs3%2Faws4_request&X-Amz-Date=20240111T133505Z&X-Amz-Expires=604800&X-Amz-SignedHeaders=host&X-Amz-Signature=ccdb4e6f980e323932d98dc5ae6ab6ba04925f7fb5fe113afa76b64df71bd92d"
        imageName:
          type: string
          readOnly: true
          description: "Название загруженной картинки"
          example: "325c0226-8acb-41ad-a214-40ff773b35bc: <2024-01-11T16:35:05.852513900> image_name.jpeg"
        createdAt:
          type: string
          format: date-time
          readOnly: true
          description: "Дата и время создания(изменения) поста"
          example: "2024-01-12 13:56"
          pattern: "yyyy-MM-dd HH:mm"
        authorName:
          type: string
          readOnly: true
          description: "Имя автора поста"
          example: "Sergey"
...
    ErrorResponse:
      type: object
      properties:
        messageError:
          type: string
          description: Error message
          example: "Error"